+++ QEDA.md
@@ -1,0 +1,18 @@
+# QEDA
+
+Quantum Electronics Design Automation - The RTL Synthesizer of Quantum Computing!
+
+## Description
+
+QEDA is an open-source, open-hardware, and open-core circuit synthesizer for the [OpenQASM Language]()! 
+
+## Prerequisites
+
+QEDA is designed to be used with SpookyIDE, our all-in-one quantum hardware development environment, you *probably* want to install [SpookyIDE](https://github.com/Spooky-Manufacturing/SpookyIDE) which will install all the pre-requisites. To run QEDA in standalone mode you will need to install the following prerequisites:
+
+* Python3.7+
+  * antlr4-python3-runtime==4.7.2 
+  * docopt**==**0.6.2
+* Kicad==5.0.2
+
+**Antlr Version is critical**.
